src/matching_service/requirements.txt
redis==4.5.5
requests==2.31.0
fastapi==0.97.0
uvicorn==0.22.0
python-dotenv==1.0.0
